% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetdocumentation.R
\docType{data}
\name{edges}
\alias{edges}
\title{edges data}
\format{
An object of class \code{data.frame} with 2016 rows and 2 columns.
}
\usage{
edges
}
\description{
An artificial data set with the edges
}
\keyword{datasets}
