% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_function.R
\name{first_function}
\alias{first_function}
\title{Sums two values}
\usage{
first_function(x, y)
}
\arguments{
\item{x}{real}

\item{y}{real}
}
\value{
real
}
\description{
Sums two values
}
\examples{
first_function(2,3)
}
