% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetdocumentation.R
\docType{data}
\name{nodes}
\alias{nodes}
\title{nodes data}
\format{
An object of class \code{data.frame} with 1249 rows and 4 columns.
}
\usage{
nodes
}
\description{
An artificial data set with the nodes
}
\keyword{datasets}
